requirements.txt
streamlit
streamlit-webrtc
openai-whisper
torch
opencv-python
mediapipe
numpy
pyttsx3
av
soundfile
